Cleaned up the generate_testcases.py file so that it's less time consuming to add new generators.
