+++ source/Generator864.py
@@ -1,0 +1,21 @@
+from source.helpers.helpers import replace_grid_field
+
+def generate() -> str:
+    grid = ["".join(["." for i in range(30)]) for j in range(30)]
+
+    replace_grid_field(grid, 0, 0, "@")
+    replace_grid_field(grid, 29, 29, "b")
+    replace_grid_field(grid, 20, 1, "a")
+    replace_grid_field(grid, 20, 2, "#")
+    replace_grid_field(grid, 21, 0, "#")
+    replace_grid_field(grid, 21, 1, "B")
+    replace_grid_field(grid, 21, 2, "#")
+    replace_grid_field(grid, 15, 15, "A")
+    replace_grid_field(grid, 14, 16, "#")
+    replace_grid_field(grid, 13, 17, "#")
+
+    test = '["' + '","'.join([row.__str__() for row in grid]) + '"]'
+
+    test2 = '["@......#.............#........",".......#######......dB........","#####..#.....#......##........",".......#...#.#..........#.....","...........#.#..........#.....","....########............#.a...","............f...........#.....","######.#E#######D########C####",".......#e...F...........#.....",".......######..........##.....",".......#..............#.......",".......#.......b....##........",".......#...........#..........",".......#.........##...........",".......#........#.............",".......#.......A..............",".......#......#...............",".......#######................","..............................","..............................","..............................","..............................","..............................","..............................","..............................","..............................","..............................","..............................","..............................",".............................c"]'
+    tests = test + "\n" + test2
+    return tests
