+++ source/helpers/helpers.py
@@ -1,9 +1,46 @@
 import random
 
+def get_testcase_random() -> [int, list, list, int, int]:
+    n, edges, succProb, start, end = get_testcase_uniform()
+    succProb = [random.random() for i in range(len(succProb))]
+    return n, edges, succProb, start, end
+
+
+def get_testcase_min() -> [int, list, list, int, int]:
+    n = 2
+    start = 0
+    end = 1
+    edges = []
+    succProb = []
+    return n, edges, succProb, start, end
+
+
+def get_testcase_uniform() -> [int, list, list, int, int]:
+    n = 10 ** 4
+    start = 0
+    end = 10 ** 4 - 1
+    edges = []
+    succProb = []
+    max_edge_count = 2 * n
+    for i in range(max_edge_count):
+        a = random.randint(start, end)
+        b = a
+        while b == a or [a, b] in edges or [b, a] in edges:
+            b = random.randint(start, end)
+        edges.append([a, b])
+        succProb.append(1)
+    return n, edges, succProb, start, end
+
+
+def replace_grid_field(grid, x, y, char):
+    grid[y] = grid[y][:x] + char + grid[y][x + 1 :]
+
+
 def get_unique_numbers(n, minVal):
     unique_numbers = list(range(minVal, minVal + n))
     random.shuffle(unique_numbers)
     return unique_numbers
+
 
 def binary_tree_with_duplicates(n, minVal, maxVal):
     '''
